Small changes while fixing up circle example.
